fix deleteProduct controller and service

+++ src/controllers/products.controller.js
@@ -92,10 +92,15 @@ const deleteProduct = async (req, res) => {
     if (isInvalidId(pid))
       return res.status(400).send({ status: "error", message: "Invalid id" });
 
-    const response = await deleteProductService(pid);
-    if (response?.status === "error") return res.status(404).send(response);
+    const product = await getProductService(pid);
+    if (!product)
+      return res
+        .status(404)
+        .send({ status: "error", message: "Product not found" });
 
-    res.send({ status: "success", message: "Product delete" });
+    const response = await deleteProductService(pid);
+
+    res.send({ status: "success", message: "Product delete", response });
   } catch (error) {
     console.log(error);
     res.status(500).send(error);
